video/remote_function/requirements.txt
flask>=3.0.2
# imutils>=0.5.4
nncf>=2.14.0
numpy>=1.23.5,<2.0.0
opencv-python-headless>=4.11.0.86
openvino-dev>=2024.6.0
onnx>=1.12.0
onnxslim
onnxruntime
# onnxruntime-gpu
sk-video>=1.1.10
ultralytics>=8.3.83
